Rename MgmtAPI to AdminAPI - part 1: classes
QubesOS/qubes-issues#853
